pyproject.toml
[project]
name = "find-stations-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "branca==0.8.1",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "dnspython==2.7.0",
    "folium==0.19.2",
    "fonttools==4.55.3",
    "geopandas==1.0.1",
    "googlemaps==4.10.0",
    "idna==3.10",
    "jinja2==3.1.5",
    "kiwisolver==1.4.8",
    "markupsafe==3.0.2",
    "matplotlib==3.10.0",
    "memory-profiler==0.61.0",
    "numpy==2.2.1",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "psutil==6.1.1",
    "pymongo==4.10.1",
    "pyogrio==0.10.0",
    "pyparsing==3.2.1",
    "pyproj==3.7.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "requests==2.32.3",
    "scipy==1.15.0",
    "shapely==2.0.6",
    "six==1.17.0",
    "tabulate==0.9.0",
    "tzdata==2024.2",
    "urllib3==2.3.0",
    "xyzservices==2024.9.0",
]
